requirements (copy).txt
certifi==2019.6.16
Click==7.0
cloudpickle==1.2.1
cycler==0.10.0
cytoolz==0.10.0
dask==2.2.0
decorator==4.4.0
Flask==1.1.1
Flask-Uploads==0.2.1
imageio==2.5.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
#####################################good till here
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
networkx==2.3 good
numpy==1.17.0-numpy 1.16.0
olefile==0.46 good
opencv-python==4.1.0.25 installed 4.1.2.30 it changed numpy to 1.24.3
Pillow==6.2.0-9.5.0
pyparsing==2.4.2 good
python-dateutil==2.8.0 good
PyWavelets==1.0.3 good
scikit-image==0.15.0  installed contourpy-1.0.7 cycler-0.11.0 fonttools-4.39.3 imageio-2.28.1 importlib-resources-5.12.0 kiwisolver-1.4.4 matplotlib-3.7.1 packaging-23.1 scikit-image-0.15.0 scipy-1.10.1 zipp-3.15.0
scikit-learn==0.21.2 scikit-learn-1.2.2
scipy==1.3.0
six==1.12.0 good
toolz==0.10.0 good
torch==1.1.0 1.4.0
torchvision==0.3.0 0.5.0
tornado==6.0.3 good
tqdm==4.32.2 good
Werkzeug==0.15.4 good
wincertstore==0.2 good
